custom video player, user profile edit, video profile edit

+++ controllers/videoControllers.js
@@ -42,7 +42,10 @@ export const postUpload = async (req, res) => {
     fileUrl: path,
     title,
     description,
+    creator: req.user.id,
   });
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   //To do : Upload and save video
   res.redirect(routes.videoDetail(newVideo.id));
 };
@@ -55,7 +58,7 @@ export const videoDetail = async (req, res) => {
   } = req;
 
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate('creator');
     res.render('videoDetail', { pageTitle: video.title, video });
   } catch (err) {
     //console.log(err);
@@ -69,7 +72,11 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (err) {
     res.redirect(routes.home);
   }
@@ -92,7 +99,12 @@ export const deleteVideo = async (req, res) => {
     params: { id },
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
     res.redirect(routes.home);
   } catch (err) {
     res.redirect(routes.home);
